# Implement DynamicMerger