… benchmark/dynamic_merger/hello.haml
… benchmark/dynamic_merger/hello.string